+++ cordapp-cpk2/src/main/java/net/corda/plugins/cpk2/CordappData.java
@@ -14,6 +14,7 @@ public class CordappData {
     private final Property<Integer> versionId;
     private final Property<String> vendor;
     private final Property<String> licence;
+    private final Property<String> cpkCordappName;
 
     @Inject
     public CordappData(@NotNull ObjectFactory objects) {
@@ -21,6 +22,7 @@ public class CordappData {
         versionId = objects.property(Integer.class);
         vendor = objects.property(String.class);
         licence = objects.property(String.class);
+        cpkCordappName = objects.property(String.class);
     }
 
     @Optional
@@ -47,8 +49,14 @@ public class CordappData {
         return licence;
     }
 
+    @Optional
+    @Input
+    public Property<String> getCpkCordappName() {
+        return cpkCordappName;
+    }
+
     boolean isEmpty() {
-        return (!name.isPresent() && !versionId.isPresent() && !vendor.isPresent() && !licence.isPresent());
+        return (!name.isPresent() && !versionId.isPresent() && !vendor.isPresent() && !licence.isPresent() && !cpkCordappName.isPresent());
     }
 
     public void name(@Nullable String value) {
@@ -65,5 +73,9 @@ public class CordappData {
 
     public void licence(@Nullable String value) {
         licence.set(value);
+    }
+
+    public void cpkCordappName(@Nullable String value) {
+        cpkCordappName.set(value);
     }
 }