+++ cordapp-cpk2/src/test/resources/with-cordapp-name/workflows/src/main/kotlin/com/example/name/workflows/ExampleFlow.kt
@@ -1,0 +1,31 @@
+@file:Suppress("unused", "PackageDirectoryMismatch")
+package com.example.name.workflows
+
+import com.example.name.contracts.ExampleState
+import net.corda.v5.application.flows.Flow
+
+class ExampleFlow(private val state: ExampleState) : Flow<String> {
+    private companion object {
+        private const val MESSAGE = "Hello, OSGi!"
+        private const val DATA = "OSGi Data"
+    }
+
+    override fun call(): String {
+        @Suppress("ObjectLiteralToLambda")
+        val innerFlow = object : Flow<String> {
+            override fun call(): String {
+                return MESSAGE
+            }
+        }
+        val lambdaFlow = Flow { DATA }
+        return "$state -> ${innerFlow.call()} AND ${lambdaFlow.call()}"
+    }
+
+    class NestedFlow : Flow<String> {
+        override fun call(): String = throw UnsupportedOperationException(this::class.java.name)
+    }
+
+    inner class InnerFlow : Flow<String> {
+        override fun call(): String = throw UnsupportedOperationException(this::class.java.name)
+    }
+}